Refactor backend config: + Refactor telnet config backend + Refactor rest config backend + Refactor web app startup
