+++ .idea/dataSources/8ff9589f-8b90-4be7-8cc6-2f814693e22d/storage_v2/_src_/database/restappdb.C2s6kQ/schema/information_schema.FNRwLQ.meta
@@ -1,2 +1,2 @@
 #n:information_schema
-!<md> [493, 0, null, null, -2147483648, -2147483648]
+!<md> [496, 0, null, null, -2147483648, -2147483648]
