+++ application/src/main/configuration/mod.rs
@@ -3,4 +3,5 @@ pub mod handle_client;
 pub mod messaging_configuration;
 pub mod persistence_configuration;
 pub mod rest_configuration;
+pub mod telnet_configuration;
 pub mod use_case_configuration;
