+++ application/src/main/mod.rs
@@ -1,3 +1,4 @@
+pub mod application_startup;
 pub mod configuration;
 pub mod controllers;
 pub mod event;
