+++ src/main.rs
@@ -1,4 +1,4 @@
-use application::main::configuration::rest_configuration::start_web_backend;
+use application::main::application_startup::start_web_backend;
 
 fn main() {
     let _ = start_web_backend();
